+++ Source/Readers/LUSequenceReader/LUSequenceWriter.cpp
@@ -70,7 +70,7 @@ void LUSequenceWriter<ElemType>::ReadLabelInfo(const wstring& vocfile,
     b = 0;
     while (!feof(vin))
     {
-        fscanf_s(vin, "%s\n", stmp, _countof(stmp));
+        fscanf_s(vin, "%s\n", stmp, (int)_countof(stmp));
         word4idx[stmp] = b;
         idx4word[b++] = stmp;
     }
